requirements.txt
torch
torch_geometric
neo4j
sentence_transformers
pandas
rdflib
pyyaml
bidict